requirements.txt
Django==4.2
djangorestframework==3.14.0
Pillow==10.4.0
python-dotenv==1.0.1
psycopg2
